refactor: move all atoms components on root `./atoms/*`
