--- src/lib/components/atoms/button/link/index.ts
@@ -1,1 +1,0 @@
-export * from "./link";
