+++ src/lib/components/atoms/checkbox/checkbox.type.ts
