+++ src/lib/components/atoms/label/label-provider.type.ts
