+++ src/lib/components/atoms/label/label.type.ts
