+++ src/lib/components/atoms/link-button/index.ts
@@ -1,0 +1,1 @@
+export * from "./link-button";
