+++ src/lib/components/atoms/switch/switch.type.ts
